+++ src/editor/core/command/CommandAdapt.ts
@@ -23,11 +23,14 @@ import {
   IGetControlValueOption,
   IGetControlValueResult,
   ISetControlExtensionOption,
+  ISetControlHighlightOption,
+  ISetControlProperties,
   ISetControlValueOption
 } from '../../interface/Control'
 import {
   IAppendElementListOption,
   IDrawImagePayload,
+  IDrawOption,
   IForceUpdateOption,
   IGetImageOption,
   IGetValueOption,
@@ -270,16 +273,32 @@ export class CommandAdapt {
       this.draw.isReadonly() || this.control.isDisabledControl()
     if (isDisabled) return
     const selection = this.range.getSelectionElementList()
-    if (!selection) return
-    selection.forEach(el => {
-      el.font = ''
-      el.color = ''
-      el.bold = false
-      el.italic = false
-      el.underline = false
-      el.strikeout = false
+    // 选区设置或设置换行处样式
+    let renderOption: IDrawOption = {}
+    let changeElementList: IElement[] = []
+    if (selection?.length) {
+      changeElementList = selection
+      renderOption = { isSetCursor: false }
+    } else {
+      const { endIndex } = this.range.getRange()
+      const elementList = this.draw.getElementList()
+      const enterElement = elementList[endIndex]
+      if (enterElement?.value === ZERO) {
+        changeElementList.push(enterElement)
+        renderOption = { curIndex: endIndex }
+      }
+    }
+    if (!changeElementList.length) return
+    changeElementList.forEach(el => {
+      delete el.size
+      delete el.font
+      delete el.color
+      delete el.bold
+      delete el.italic
+      delete el.underline
+      delete el.strikeout
     })
-    this.draw.render({ isSetCursor: false })
+    this.draw.render(renderOption)
   }
 
   public font(payload: string) {
@@ -287,11 +306,20 @@ export class CommandAdapt {
       this.draw.isReadonly() || this.control.isDisabledControl()
     if (isDisabled) return
     const selection = this.range.getSelectionElementList()
-    if (!selection) return
-    selection.forEach(el => {
-      el.font = payload
-    })
-    this.draw.render({ isSetCursor: false })
+    if (selection?.length) {
+      selection.forEach(el => {
+        el.font = payload
+      })
+      this.draw.render({ isSetCursor: false })
+    } else {
+      const { endIndex } = this.range.getRange()
+      const elementList = this.draw.getElementList()
+      const enterElement = elementList[endIndex]
+      if (enterElement?.value === ZERO) {
+        enterElement.font = payload
+        this.draw.render({ curIndex: endIndex, isCompute: false })
+      }
+    }
   }
 
   public size(payload: number) {
@@ -300,10 +328,25 @@ export class CommandAdapt {
     const isDisabled =
       this.draw.isReadonly() || this.control.isDisabledControl()
     if (isDisabled) return
+    // 选区设置或设置换行处样式
+    let renderOption: IDrawOption = {}
+    let changeElementList: IElement[] = []
     const selection = this.range.getTextLikeSelectionElementList()
-    if (!selection || !selection.length) return
+    if (selection?.length) {
+      changeElementList = selection
+      renderOption = { isSetCursor: false }
+    } else {
+      const { endIndex } = this.range.getRange()
+      const elementList = this.draw.getElementList()
+      const enterElement = elementList[endIndex]
+      if (enterElement?.value === ZERO) {
+        changeElementList.push(enterElement)
+        renderOption = { curIndex: endIndex }
+      }
+    }
+    if (!changeElementList.length) return
     let isExistUpdate = false
-    selection.forEach(el => {
+    changeElementList.forEach(el => {
       if (
         (!el.size && payload === defaultSize) ||
         (el.size && el.size === payload)
@@ -314,7 +357,7 @@ export class CommandAdapt {
       isExistUpdate = true
     })
     if (isExistUpdate) {
-      this.draw.render({ isSetCursor: false })
+      this.draw.render(renderOption)
     }
   }
 
@@ -323,10 +366,25 @@ export class CommandAdapt {
       this.draw.isReadonly() || this.control.isDisabledControl()
     if (isDisabled) return
     const selection = this.range.getTextLikeSelectionElementList()
-    if (!selection || !selection.length) return
+    // 选区设置或设置换行处样式
+    let renderOption: IDrawOption = {}
+    let changeElementList: IElement[] = []
+    if (selection?.length) {
+      changeElementList = selection
+      renderOption = { isSetCursor: false }
+    } else {
+      const { endIndex } = this.range.getRange()
+      const elementList = this.draw.getElementList()
+      const enterElement = elementList[endIndex]
+      if (enterElement?.value === ZERO) {
+        changeElementList.push(enterElement)
+        renderOption = { curIndex: endIndex }
+      }
+    }
+    if (!changeElementList.length) return
     const { defaultSize, maxSize } = this.options
     let isExistUpdate = false
-    selection.forEach(el => {
+    changeElementList.forEach(el => {
       if (!el.size) {
         el.size = defaultSize
       }
@@ -339,7 +397,7 @@ export class CommandAdapt {
       isExistUpdate = true
     })
     if (isExistUpdate) {
-      this.draw.render({ isSetCursor: false })
+      this.draw.render(renderOption)
     }
   }
 
@@ -348,10 +406,25 @@ export class CommandAdapt {
       this.draw.isReadonly() || this.control.isDisabledControl()
     if (isDisabled) return
     const selection = this.range.getTextLikeSelectionElementList()
-    if (!selection || !selection.length) return
+    // 选区设置或设置换行处样式
+    let renderOption: IDrawOption = {}
+    let changeElementList: IElement[] = []
+    if (selection?.length) {
+      changeElementList = selection
+      renderOption = { isSetCursor: false }
+    } else {
+      const { endIndex } = this.range.getRange()
+      const elementList = this.draw.getElementList()
+      const enterElement = elementList[endIndex]
+      if (enterElement?.value === ZERO) {
+        changeElementList.push(enterElement)
+        renderOption = { curIndex: endIndex }
+      }
+    }
+    if (!changeElementList.length) return
     const { defaultSize, minSize } = this.options
     let isExistUpdate = false
-    selection.forEach(el => {
+    changeElementList.forEach(el => {
       if (!el.size) {
         el.size = defaultSize
       }
@@ -364,7 +437,7 @@ export class CommandAdapt {
       isExistUpdate = true
     })
     if (isExistUpdate) {
-      this.draw.render({ isSetCursor: false })
+      this.draw.render(renderOption)
     }
   }
 
@@ -373,12 +446,21 @@ export class CommandAdapt {
       this.draw.isReadonly() || this.control.isDisabledControl()
     if (isDisabled) return
     const selection = this.range.getSelectionElementList()
-    if (!selection) return
-    const noBoldIndex = selection.findIndex(s => !s.bold)
-    selection.forEach(el => {
-      el.bold = !!~noBoldIndex
-    })
-    this.draw.render({ isSetCursor: false })
+    if (selection?.length) {
+      const noBoldIndex = selection.findIndex(s => !s.bold)
+      selection.forEach(el => {
+        el.bold = !!~noBoldIndex
+      })
+      this.draw.render({ isSetCursor: false })
+    } else {
+      const { endIndex } = this.range.getRange()
+      const elementList = this.draw.getElementList()
+      const enterElement = elementList[endIndex]
+      if (enterElement?.value === ZERO) {
+        enterElement.bold = !enterElement.bold
+        this.draw.render({ curIndex: endIndex, isCompute: false })
+      }
+    }
   }
 
   public italic() {
@@ -386,12 +468,21 @@ export class CommandAdapt {
       this.draw.isReadonly() || this.control.isDisabledControl()
     if (isDisabled) return
     const selection = this.range.getSelectionElementList()
-    if (!selection) return
-    const noItalicIndex = selection.findIndex(s => !s.italic)
-    selection.forEach(el => {
-      el.italic = !!~noItalicIndex
-    })
-    this.draw.render({ isSetCursor: false })
+    if (selection?.length) {
+      const noItalicIndex = selection.findIndex(s => !s.italic)
+      selection.forEach(el => {
+        el.italic = !!~noItalicIndex
+      })
+      this.draw.render({ isSetCursor: false })
+    } else {
+      const { endIndex } = this.range.getRange()
+      const elementList = this.draw.getElementList()
+      const enterElement = elementList[endIndex]
+      if (enterElement?.value === ZERO) {
+        enterElement.italic = !enterElement.italic
+        this.draw.render({ curIndex: endIndex, isCompute: false })
+      }
+    }
   }
 
   public underline() {
@@ -399,15 +490,24 @@ export class CommandAdapt {
       this.draw.isReadonly() || this.control.isDisabledControl()
     if (isDisabled) return
     const selection = this.range.getSelectionElementList()
-    if (!selection) return
-    const noUnderlineIndex = selection.findIndex(s => !s.underline)
-    selection.forEach(el => {
-      el.underline = !!~noUnderlineIndex
-    })
-    this.draw.render({
-      isSetCursor: false,
-      isCompute: false
-    })
+    if (selection?.length) {
+      const noUnderlineIndex = selection.findIndex(s => !s.underline)
+      selection.forEach(el => {
+        el.underline = !!~noUnderlineIndex
+      })
+      this.draw.render({
+        isSetCursor: false,
+        isCompute: false
+      })
+    } else {
+      const { endIndex } = this.range.getRange()
+      const elementList = this.draw.getElementList()
+      const enterElement = elementList[endIndex]
+      if (enterElement?.value === ZERO) {
+        enterElement.underline = !enterElement.underline
+        this.draw.render({ curIndex: endIndex, isCompute: false })
+      }
+    }
   }
 
   public strikeout() {
@@ -415,15 +515,24 @@ export class CommandAdapt {
       this.draw.isReadonly() || this.control.isDisabledControl()
     if (isDisabled) return
     const selection = this.range.getSelectionElementList()
-    if (!selection) return
-    const noStrikeoutIndex = selection.findIndex(s => !s.strikeout)
-    selection.forEach(el => {
-      el.strikeout = !!~noStrikeoutIndex
-    })
-    this.draw.render({
-      isSetCursor: false,
-      isCompute: false
-    })
+    if (selection?.length) {
+      const noStrikeoutIndex = selection.findIndex(s => !s.strikeout)
+      selection.forEach(el => {
+        el.strikeout = !!~noStrikeoutIndex
+      })
+      this.draw.render({
+        isSetCursor: false,
+        isCompute: false
+      })
+    } else {
+      const { endIndex } = this.range.getRange()
+      const elementList = this.draw.getElementList()
+      const enterElement = elementList[endIndex]
+      if (enterElement?.value === ZERO) {
+        enterElement.strikeout = !enterElement.strikeout
+        this.draw.render({ curIndex: endIndex, isCompute: false })
+      }
+    }
   }
 
   public superscript() {
@@ -491,14 +600,23 @@ export class CommandAdapt {
       this.draw.isReadonly() || this.control.isDisabledControl()
     if (isDisabled) return
     const selection = this.range.getSelectionElementList()
-    if (!selection) return
-    selection.forEach(el => {
-      el.color = payload
-    })
-    this.draw.render({
-      isSetCursor: false,
-      isCompute: false
-    })
+    if (selection?.length) {
+      selection.forEach(el => {
+        el.color = payload
+      })
+      this.draw.render({
+        isSetCursor: false,
+        isCompute: false
+      })
+    } else {
+      const { endIndex } = this.range.getRange()
+      const elementList = this.draw.getElementList()
+      const enterElement = elementList[endIndex]
+      if (enterElement?.value === ZERO) {
+        enterElement.color = payload
+        this.draw.render({ curIndex: endIndex, isCompute: false })
+      }
+    }
   }
 
   public highlight(payload: string) {
@@ -506,14 +624,23 @@ export class CommandAdapt {
       this.draw.isReadonly() || this.control.isDisabledControl()
     if (isDisabled) return
     const selection = this.range.getSelectionElementList()
-    if (!selection) return
-    selection.forEach(el => {
-      el.highlight = payload
-    })
-    this.draw.render({
-      isSetCursor: false,
-      isCompute: false
-    })
+    if (selection?.length) {
+      selection.forEach(el => {
+        el.highlight = payload
+      })
+      this.draw.render({
+        isSetCursor: false,
+        isCompute: false
+      })
+    } else {
+      const { endIndex } = this.range.getRange()
+      const elementList = this.draw.getElementList()
+      const enterElement = elementList[endIndex]
+      if (enterElement?.value === ZERO) {
+        enterElement.highlight = payload
+        this.draw.render({ curIndex: endIndex, isCompute: false })
+      }
+    }
   }
 
   public title(payload: TitleLevel | null) {
@@ -558,34 +685,7 @@ export class CommandAdapt {
   public list(listType: ListType | null, listStyle?: ListStyle) {
     const isReadonly = this.draw.isReadonly()
     if (isReadonly) return
-    const { startIndex, endIndex } = this.range.getRange()
-    if (!~startIndex && !~endIndex) return
-    // 需要改变的元素列表
-    const changeElementList = this.range.getRangeParagraphElementList()
-    if (!changeElementList || !changeElementList.length) return
-    // 如果包含列表则设置为取消列表
-    const isUnsetList = changeElementList.find(
-      el => el.listType === listType && el.listStyle === listStyle
-    )
-    // 设置值
-    const listId = getUUID()
-    changeElementList.forEach(el => {
-      if (!isUnsetList && listType) {
-        el.listId = listId
-        el.listType = listType
-        el.listStyle = listStyle
-      } else {
-        if (el.listId) {
-          delete el.listId
-          delete el.listType
-          delete el.listStyle
-        }
-      }
-    })
-    // 光标定位
-    const isSetCursor = startIndex === endIndex
-    const curIndex = isSetCursor ? endIndex : startIndex
-    this.draw.render({ curIndex, isSetCursor })
+    this.draw.getListParticle().setList(listType, listStyle)
   }
 
   public rowFlex(payload: RowFlex) {
@@ -1619,29 +1719,18 @@ export class CommandAdapt {
     const isDisabled =
       this.draw.isReadonly() || this.control.isDisabledControl()
     if (isDisabled) return
-    const activeControl = this.control.getActiveControl()
-    if (activeControl) return
     const { startIndex, endIndex } = this.range.getRange()
     if (!~startIndex && !~endIndex) return
-    const elementList = this.draw.getElementList()
     const { value, width, height } = payload
-    const element: IElement = {
-      value,
-      width,
-      height,
-      id: getUUID(),
-      type: ElementType.IMAGE
-    }
-    const curIndex = startIndex + 1
-    formatElementContext(elementList, [element], startIndex)
-    this.draw.spliceElementList(
-      elementList,
-      curIndex,
-      startIndex === endIndex ? 0 : endIndex - startIndex,
-      element
-    )
-    this.range.setRange(curIndex, curIndex)
-    this.draw.render({ curIndex })
+    this.draw.insertElementList([
+      {
+        value,
+        width,
+        height,
+        id: getUUID(),
+        type: ElementType.IMAGE
+      }
+    ])
   }
 
   public search(payload: string | null) {
@@ -1945,13 +2034,16 @@ export class CommandAdapt {
         height: lineHeight
       })
     }
+    // 区域信息
+    const zone = this.draw.getZone().getZone()
     return deepClone({
       isCollapsed,
       startElement,
       endElement,
       startPageNo,
       endPageNo,
-      rangeRects
+      rangeRects,
+      zone
     })
   }
 
@@ -2195,6 +2287,16 @@ export class CommandAdapt {
     const isReadonly = this.draw.isReadonly()
     if (isReadonly) return
     this.draw.getControl().setExtensionByConceptId(payload)
+  }
+
+  public setControlProperties(payload: ISetControlProperties) {
+    const isReadonly = this.draw.isReadonly()
+    if (isReadonly) return
+    this.draw.getControl().setPropertiesByConceptId(payload)
+  }
+
+  public setControlHighlight(payload: ISetControlHighlightOption) {
+    this.draw.getControl().setHighlightList(payload)
   }
 
   public getContainer(): HTMLDivElement {