+++ src/editor/core/position/Position.ts
@@ -229,7 +229,7 @@ export class Position {
     let startRowIndex = 0
     for (let i = 0; i < pageRowList.length; i++) {
       const rowList = pageRowList[i]
-      const startIndex = rowList[0].startIndex
+      const startIndex = rowList[0]?.startIndex
       this.computePageRowPosition({
         positionList: this.positionList,
         rowList,
@@ -288,7 +288,7 @@ export class Position {
       positionList = this.getOriginalPositionList()
     }
     const zoneManager = this.draw.getZone()
-    const curPageNo = this.draw.getPageNo()
+    const curPageNo = payload.pageNo ?? this.draw.getPageNo()
     const isMainActive = zoneManager.isMainActive()
     const positionNo = isMainActive ? curPageNo : 0
     for (let j = 0; j < positionList.length; j++) {